extract ShaderHoverProvider
