+++ src/hlsl/hoverProvider.ts
@@ -1,79 +1,15 @@
 'use strict';
 
-import { HoverProvider, Hover, SymbolInformation, SymbolKind, MarkdownString, MarkedString, TextDocument, CancellationToken, Range, Position, Uri, ViewColumn, Disposable, commands, window, workspace, WebviewPanel } from 'vscode';
-import { HTML_TEMPLATE } from './html';
+import { Hover, SymbolKind, MarkdownString, MarkedString, TextDocument, CancellationToken, Range, Position, workspace } from 'vscode';
 import hlslGlobals = require('./hlslGlobals');
-import { https } from 'follow-redirects';
-import { JSDOM } from 'jsdom';
+import { linkToMarkdownString, textToMarkedString } from '../markdown';
+import { ShaderHoverProvider } from './ShaderHoverProvider';
 
-export function textToMarkedString(text: string): MarkedString {
-	return text.replace(/[\\`*_{}[\]()#+\-.!]/g, '\\$&'); // escape markdown syntax tokens: http://daringfireball.net/projects/markdown/syntax#backslash
-}
-
-export function linkToMarkdownString(linkUrl: string): MarkdownString {
-    if (linkUrl === undefined || linkUrl === '') {
-        return;
-    }
-
-    let link = new MarkdownString('[HLSL documentation][1]\n\n[1]: ');
-    let openDocOnSide = workspace.getConfiguration('hlsl').get<boolean>('openDocOnSide', false);
-    if (openDocOnSide) {
-        link.appendText(encodeURI( 'command:shader.openLink?' + JSON.stringify([linkUrl, true])));
-    } else {
-        link.appendText(linkUrl);
-    }
-    link.isTrusted = true;
-    return link;
-}
-
-export default class HLSLHoverProvider implements HoverProvider {
-
-    private _subscriptions: Disposable[] = [];
-    private _panel: WebviewPanel = null;
-
-    private getSymbols(document: TextDocument): Thenable<SymbolInformation[]> {
-        return commands.executeCommand<SymbolInformation[]>('vscode.executeDocumentSymbolProvider', document.uri);
-    }
-
+export default class HLSLHoverProvider extends ShaderHoverProvider {
     constructor() {
-        this._subscriptions.push( commands.registerCommand('shader.openLink', (link: string, newWindow: boolean) => {
-            if (!this._panel) {
-                this._panel = window.createWebviewPanel(
-                    'hlsldoc',
-                    'HLSL Documentation',
-                    newWindow ? ViewColumn.Two : ViewColumn.Active,
-                    {
-                        // Enable scripts in the webview
-                        enableScripts: true
-                    }
-                );
-
-                this._panel.onDidDispose( () => {
-                    this._panel = null;
-                });
-
-                this._panel.webview.onDidReceiveMessage(
-                    message => {
-                        switch (message.command) {
-                            case 'clickLink':
-                                commands.executeCommand('shader.openLink', message.text);
-                                return;
-                        }
-                    }
-                );
-            }
-            this._panel.reveal();
-            // And set its HTML content
-            getWebviewContent(link).then(html => this._panel.webview.html = html);
-        }));
-
+        super('hlsl');
     }
-
-    dispose() {
-        this._subscriptions.forEach(s => {s.dispose()});
-    }
-
-
+    
     public async provideHover(document: TextDocument, position: Position, token: CancellationToken): Promise<Hover> {
         
         let enable = workspace.getConfiguration('hlsl').get<boolean>('suggest.basic', true);
@@ -192,45 +128,4 @@ export default class HLSLHoverProvider implements HoverProvider {
     } 
 }
 
-function getWebviewContent(link: string): Promise<string> {
-    const uri = Uri.parse(link);
-    return new Promise<string>((resolve, reject) => {
-        let request = https.request({
-            host: uri.authority,
-            path: uri.path,
-            rejectUnauthorized: workspace.getConfiguration().get("http.proxyStrictSSL", true)
-        }, (response) => {
-            if (response.statusCode == 301 || response.statusCode == 302)
-                return resolve(response.headers.location);
-            if (response.statusCode != 200)
-                return resolve(response.statusCode.toString());
-            let html = '';
-            response.on('data', (data) => { html += data.toString(); });
-            response.on('end', () => {
-                const dom = new JSDOM(html);
-                let topic = '';
-                let node = dom.window.document.querySelector('.content');
-                if (node) {
-                    let num = node.getElementsByTagName('a').length;
-                    for (let i = 0; i < num; ++i) {
-                        const href = node.getElementsByTagName('a')[i].href;
-                        const fulllink = new dom.window.URL(href, uri.toString()).href
-                        node.getElementsByTagName('a')[i].href = '#';
-                        node.getElementsByTagName('a')[i].setAttribute('onclick', `clickLink('${fulllink}')`)
-                    }
-                    node.querySelector('.metadata.page-metadata')?.remove();
-                    node.querySelector('#center-doc-outline')?.remove();
-                    topic = node.outerHTML;
 
-                } else {
-                    let link = uri.with({ scheme: 'https' }).toString();
-                    topic = `<a href="${link}">No topic found, click to follow link</a>`;
-                }
-                resolve(HTML_TEMPLATE.replace('{0}', topic));
-            });
-            response.on('error', (error) => { console.log(error); });
-        });
-        request.on('error', (error) => { console.log(error) });
-        request.end();
-    });
-}
